更改变量类型

+++ USERLIB/moduleconfig.cpp
@@ -1,8 +1,8 @@
 #include "moduleconfig.h"
 
 ModuleConfig ModuleConfig::_instance;
-size_t ModuleConfig::flash_write_16bit_len_ = sizeof(struct ModuleConfigData)/sizeof(uint16_t);
-size_t ModuleConfig::flash_free_page_base_addr_ = 0x08007C00;
+uint32_t ModuleConfig::flash_write_16bit_len_ = sizeof(struct ModuleConfigData)/sizeof(uint16_t);
+uint32_t ModuleConfig::flash_free_page_base_addr_ = 0x08007C00;
 
 bool ModuleConfig::ModuleConfigInit() {
     if(config_init_state_ == false){
